requirements.txt
torch 
torchvision
ollama
transformers
python-dotenv
fast-mcp
mcp[cli]
mcp
fastmcp
langchain-google-genai